creating load_shape class
